Rename Project => V1Project
