+++ src/components/V1Project/Pay/index.tsx
@@ -1,8 +1,8 @@
 import { Button, Tooltip } from 'antd'
 import { t, Trans } from '@lingui/macro'
 
-import ConfirmPayOwnerModal from 'components/Project/modals/ConfirmPayOwnerModal'
-import PayWarningModal from 'components/Project/modals/PayWarningModal'
+import ConfirmPayOwnerModal from 'components/V1Project/modals/ConfirmPayOwnerModal'
+import PayWarningModal from 'components/V1Project/modals/PayWarningModal'
 import InputAccessoryButton from 'components/shared/InputAccessoryButton'
 import FormattedNumberInput from 'components/shared/inputs/FormattedNumberInput'
 import { V1ProjectContext } from 'contexts/v1/projectContext'
