+++ src/components/V1Project/Rewards/index.tsx
@@ -1,9 +1,9 @@
 import { BigNumber } from '@ethersproject/bignumber'
 import { t, Trans } from '@lingui/macro'
 import { Button, Descriptions, Modal, Space, Statistic, Tooltip } from 'antd'
-import ConfirmUnstakeTokensModal from 'components/Project/modals/ConfirmUnstakeTokensModal'
-import ParticipantsModal from 'components/Project/modals/ParticipantsModal'
-import RedeemModal from 'components/Project/modals/RedeemModal'
+import ConfirmUnstakeTokensModal from 'components/V1Project/modals/ConfirmUnstakeTokensModal'
+import ParticipantsModal from 'components/V1Project/modals/ParticipantsModal'
+import RedeemModal from 'components/V1Project/modals/RedeemModal'
 import FormattedAddress from 'components/shared/FormattedAddress'
 import { NetworkContext } from 'contexts/networkContext'
 import { V1ProjectContext } from 'contexts/v1/projectContext'
